allow gz_lib-ng (without compat)
I have a tool where, for stupid zlib linking reasons, I need to use flate2's zlib-ng and not zlib-ng-compat.  This patch bumps the deps and adds that optional feature to niffler.

To allow building my other crate, I've created a fork crate (`niffler-temp`) but I can revert back to this once this is merged and a new version put on crates.io.

Thanks!